* [CHD-886] Enclosing a URL in parenthesis within reply adds the closing ) to the link
* [CHD-817] [Forums] No clear indicator that the checkbox next to a forum is used to delete it

* [CHD-888] [Licenses] The 'buy license' links are not pointing to the actual buy page on the website

* [CHD-776] [DEMO MODE] Certain 'helpdesk setup' tabs can still be modified

git-svn-id: http://svn.webgroupmedia.com/cerb4/cerb4@797 58cff151-b722-0410-b014-f78a459f8f02

+++ plugins/cerberusweb.webapi/api/App.php
@@ -99,6 +99,11 @@ class ChWebApiConfigTab extends Extension_ConfigTab {
 		@$add_nickname = DevblocksPlatform::importGPC($_REQUEST['add_nickname'],'string');
 		@$deletes = DevblocksPlatform::importGPC($_REQUEST['deletes'],'array',array());
 		
+		if(DEMO_MODE) {
+			DevblocksPlatform::setHttpResponse(new DevblocksHttpResponse(array('config','webapi')));
+			return;
+		}		
+		
 		// Deletes
 		if(!empty($deletes)) {
 			DAO_WebapiKey::delete($deletes);
